# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/bot": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves a list of all bots for user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "Get a list of all bots",
                "responses": {
                    "200": {
                        "description": "List of bots retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/bot.BotStructArrayResponse"
                        }
                    },
                    "400": {
                        "description": "Error getting bots",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/bot/": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new bot with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "Create a new bot",
                "parameters": [
                    {
                        "description": "Bot request body",
                        "name": "botRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bot.BotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bot created successfully",
                        "schema": {
                            "$ref": "#/definitions/bot.BotStructResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/bot/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an existing bot with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "Update an existing bot",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Bot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Bot request body",
                        "name": "botRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bot.BotUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bot updated successfully",
                        "schema": {
                            "$ref": "#/definitions/bot.BotStructResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bot.BotRequest": {
            "type": "object",
            "required": [
                "name",
                "platform"
            ],
            "properties": {
                "botLink": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "bot.BotResponse": {
            "type": "object",
            "required": [
                "name",
                "platform"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "botLink": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "bot.BotStructArrayResponse": {
            "type": "object",
            "properties": {
                "bots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bot.BotResponse"
                    }
                }
            }
        },
        "bot.BotStructResponse": {
            "type": "object",
            "properties": {
                "bot": {
                    "$ref": "#/definitions/bot.BotResponse"
                }
            }
        },
        "bot.BotUpdateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}